minor use of size_t as sizes

+++ src/tools/readbuf.h
@@ -1,6 +1,8 @@
 #ifndef READBUF_H
 #define READBUF_H
 
+#include <stddef.h>
+
 // Implement a read buffer, which is intended to be used for read/ReadFile
 // operations. Before doing a read of <n> bytes, call <prepare_read> with <n>
 // as argument. After the read, register the number of bytes read using
@@ -9,22 +11,22 @@
 
 typedef struct {
   char* data;
-  int len;
-  int capacity;
+  size_t len;
+  size_t capacity;
 } t_readbuf, *preadbuf;
 
 // return a read buf of initial capacity <capacity>
-preadbuf init_readbuf(int capacity);
+preadbuf init_readbuf(size_t capacity);
 
 // return a pointer to a memory region which is unused and can act as a buffer
 // for a read operation reading up to size bytes
-char* prepare_read(preadbuf b, int size);
+char* prepare_read(preadbuf b, size_t size);
 
 //notify the buffer that <size> bytes have been read
-void have_read(preadbuf b, int size);
+void have_read(preadbuf b, size_t size);
 
 // allow the readbuf to delete the first <size> byte of the buffer
-void have_taken(preadbuf b, int size);
+void have_taken(preadbuf b, size_t size);
 
 // allow the readbuf to all of the buffer
 void clear_readbuf(preadbuf b);
